Improvement

+++ run.sh
@@ -2,6 +2,13 @@
 
 HERE="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 
+if [ "$1" = 'k8s' ]; then # For easy access through the Docker image without cloning the repo
+    cat "$HERE"/k8s/prep.sh
+    echo ""
+    echo "# Put the above script into a file and run it on your K8s control plane. Make the resulting files accessible to Logtfy."
+    exit
+fi
+
 trap "if [ -f "$HERE"/onExit.sh ]; then bash "$HERE"/onExit.sh; else bash "$HERE"/onExit.default.sh; fi; trap - SIGTERM && kill -- -\$\$" EXIT
 
 mkdir -p /tmp/logtfy # Semi-persistent storage used by some modules
